requirements.txt
python-dotenv
langchain_google_genai
langchain_core
langchain_community
langchain_pinecone
uuid
fastapi[standard]
pymongo
cloudinary
bs4